+++ burn-autodiff/src/checkpoint/tests.rs
@@ -1,200 +1,18 @@
-use std::{any::Any, collections::HashMap, marker::PhantomData};
-
-use burn_tensor::{backend::Backend, Tensor};
-
-use super::{NodeID, NodeRef};
-
-trait RetroForward {
-    fn forward(&self, states: &mut InnerStates);
-}
-
-#[derive(new)]
-pub struct RetroLeaf<B: Backend, const D: usize> {
-    out: NodeID,
-    tensor: Tensor<B, D>, // maybe remove that state and just have retroleaves as always computed
-}
-
-impl<B: Backend, const D: usize> RetroForward for RetroLeaf<B, D> {
-    fn forward(&self, states: &mut InnerStates) {
-        states.insert(
-            self.out.clone(),
-            State::Computed {
-                state_content: Box::new(self.tensor.clone()), // must not clone tensor
-                n_required: 1,                                // TODO arbitrary for now
-            },
-        );
-    }
-}
-
-#[derive(new)]
-pub struct RetroDiv<B, const D: usize> {
-    lhs: NodeID,
-    rhs: NodeID,
-    out: NodeID,
-    _backend: PhantomData<B>,
-}
-
-impl<B: Backend, const D: usize> RetroForward for RetroDiv<B, D> {
-    fn forward(&self, states: &mut InnerStates) {
-        // We assume hashmap filled with parents
-        let lhs: B::FloatTensorPrimitive<D> = states
-            .get(&self.lhs)
-            .get_state_content()
-            .downcast_ref::<Tensor<B, D>>()
-            .unwrap()
-            .clone()
-            .into_primitive();
-
-        let rhs: B::FloatTensorPrimitive<D> = states
-            .get(&self.rhs) // TODO get_mut because change num_required -=1
-            .get_state_content()
-            .downcast_ref::<Tensor<B, D>>()
-            .unwrap()
-            .clone()
-            .into_primitive();
-
-        let out: Tensor<B, D> = Tensor::<B, D>::from_primitive(B::float_div(lhs, rhs));
-
-        states.insert(
-            self.out.clone(),
-            State::Computed {
-                state_content: Box::new(out),
-                n_required: 1, // TODO lazy's
-            },
-        );
-    }
-}
-
-type StateContent = Box<dyn Any + Send + Sync>;
-
-#[derive(Debug)]
-enum State {
-    // Weird nomenclature. Isn't it more lazy to not re-compute?
-    Lazy {
-        node_id: NodeID, // whose forward is required to compute state (is it needed, as States has it as the key)
-        n_required: usize, // how many times it's used (has counter += and -=)
-    },
-    Computed {
-        state_content: StateContent,
-        n_required: usize,
-    },
-}
-
-impl State {
-    fn get_state_content(&self) -> &StateContent {
-        match self {
-            State::Lazy {
-                node_id: _,
-                n_required: _,
-            } => unreachable!("A child has been called before its parents"),
-            State::Computed {
-                state_content,
-                n_required: _,
-            } => state_content,
-        }
-    }
-}
-
-#[derive(new, Default)]
-struct RetroForwards {
-    map: HashMap<NodeID, Box<dyn RetroForward>>,
-}
-
-impl RetroForwards {
-    pub fn forward(&self, node_id: &NodeID, inner_states: &mut InnerStates) {
-        if let State::Lazy {
-            node_id,
-            n_required,
-        } = inner_states.get(node_id)
-        {
-            self.map.get(node_id).unwrap().forward(inner_states);
-        }
-    }
-
-    pub fn insert(&mut self, node_id: NodeID, retro_forward: Box<dyn RetroForward>) {
-        self.map.insert(node_id, retro_forward);
-    }
-}
-
-// wrapper to keep track of n_requiered. if zero remove and give ownership. in the REAL forward +=1 n_required
-#[derive(new, Default)]
-struct InnerStates {
-    map: HashMap<NodeID, State>,
-}
-
-impl InnerStates {
-    pub fn get(&self, node_id: &NodeID) -> &State {
-        self.map.get(node_id).unwrap()
-    }
-
-    pub fn insert(&mut self, node_id: NodeID, state: State) {
-        self.map.insert(node_id, state);
-    }
-}
-
-#[derive(new)]
-struct States {
-    inner_states: InnerStates,
-    retro_forwards: RetroForwards,
-    nodes: HashMap<NodeID, NodeRef>,
-}
-
-impl States {
-    pub fn get(&mut self, node_id: NodeID) -> &StateContent {
-        self.topological_sort(node_id.clone())
-            .iter()
-            .for_each(|node| self.retro_forwards.forward(&node, &mut self.inner_states));
-
-        self.inner_states.get(&node_id).get_state_content()
-    }
-
-    fn topological_sort(&self, node_id: NodeID) -> Vec<NodeID> {
-        match self.inner_states.get(&node_id) {
-            State::Lazy {
-                node_id: _,
-                n_required: _,
-            } => {
-                let mut sorted = Vec::new();
-                for parent_node in self.parents(node_id.clone()) {
-                    sorted.extend(self.topological_sort(parent_node));
-                }
-                sorted.push(node_id);
-                sorted
-            }
-            State::Computed {
-                state_content: _,
-                n_required: _,
-            } => vec![node_id],
-        }
-    }
-
-    fn parents(&self, node_id: NodeID) -> Vec<NodeID> {
-        self.nodes.get(&node_id).unwrap().parents.clone()
-    }
-}
-
-// progressive transformation:
-// all operations saved a custom state for their backward (if necessary)
-// this does not mean they are able to perform their forward
-// we must leave them the ability of computing their backward from their state
-// but not assume it means they have what they need for the retro forward
-// conclusion: we do not withdraw the original state, but it's useless for checkpointing, therefore by default all operations are lazy (memory bound)
-// some ops will do their own backward with their state, some will ask their ancestors to retroforward, but there's no link.
-// have performance tests to decide if we keep both or uniformize
-
-// during a backward,
-// an operation may need its original inputs
-// it will ask the states struct's get. twice if it has two inputs
-// OR just once, asking for its own node id, should work just as well
-
 #[cfg(test)]
 mod tests {
 
     use std::sync::Arc;
 
-    use burn_tensor::Data;
+    use burn_tensor::{backend::Backend, Data, Tensor};
 
-    use crate::graph::{Node, Requirement};
+    use crate::{
+        checkpoint::{
+            retro::{RetroDiv, RetroLeaf},
+            state::{State, StateContent},
+            states::{InnerStates, NodeTree, RetroForwards, States},
+        },
+        graph::{Node, NodeID, NodeRef, Requirement},
+    };
 
     use super::*;
 
@@ -230,16 +48,16 @@ mod tests {
     fn make_leaves<B: Backend>(
         device: &B::Device,
         ids: [NodeID; 4],
-    ) -> (InnerStates, RetroForwards, HashMap<NodeID, NodeRef>) {
+    ) -> (InnerStates, RetroForwards, NodeTree) {
         let mut retro_forwards = RetroForwards::default();
-        let mut nodes = HashMap::new();
+        let mut node_tree = NodeTree::default();
         let mut inner_states = InnerStates::default();
 
         let mut make_leaf = |id: NodeID, t: Tensor<B, 2>| {
             let n: NodeRef = Arc::new(Node::new(Vec::new(), 0, id.clone(), Requirement::Grad));
             let retro_leaf = RetroLeaf::new(id.clone(), t);
             retro_forwards.insert(id.clone(), Box::new(retro_leaf));
-            nodes.insert(id.clone(), n);
+            node_tree.insert(id.clone(), n);
             insert_lazy(id, &mut inner_states);
         };
 
@@ -283,7 +101,7 @@ mod tests {
             .require_grad(),
         );
 
-        (inner_states, retro_forwards, nodes)
+        (inner_states, retro_forwards, node_tree)
     }
 
     fn div_lazy_tree<B: Backend>(device: &B::Device, ids: [NodeID; 7]) -> States {
